Split leaderboard into lb and table, refactor
