+++ src/components/molecules/Table/index.ts
@@ -1,0 +1,1 @@
+export { default as Table } from './Table';
