+++ src/components/organisms/Leaderboard/index.ts
